Split out mongo code
